requirements.txt
aioEasyPillow==0.0.3
aiohttp==3.9.3
asyncache==0.3.1
Babel==2.14.0
beautifulsoup4==4.12.3
cachetools==5.3.3
cchardet==2.1.7
cloudscraper==1.2.71
discord.py==2.3.2
howlongtobeatpy==1.0.5
lxml==5.1.0
numpy==1.26.4
Pillow==10.3.0
psnawp_api==0.0.1
psycopg2==2.9.9
Requests==2.31.0
selenium==4.19.0
